Add Ruby examples
Include all strategies tested in other languages. For comment out
and early return the strategy needed to be modified in order to
make it work but the idea still works.

Unlike the other languages invisible functions do work in Ruby since
any unicode character can be put in a method name.